kernel: mtd: spi-nor: allow use of mixed erasesizes (E.g. 64K & 4K)
Allow SPI NOR code to be configured to use the multiple erase-regions
code path for a uniform erase-region device. This code path allows an
erase request to select and run a list of erase operations (including 4K
erases where supported by the device, and needed by the request). The
multi erase-regions code path does additional checks before the erase
starts, therefore may take slightly longer before the erase runs on device.

Kernel config options:
Introduces:
CONFIG_MTD_SPI_NOR_USE_VARIABLE_ERASE=y
Used with:
CONFIG_MTD_SPI_NOR_USE_4K_SECTORS=n

Previous users of partial or 4K erase in Openwrt:
- Mikrotik routerboot bootloader configuration
  target/linux/generic/files/drivers/platform/mikrotik/rb_softconfig.c
  Will work with this patch, with the 4K_SECTORS test removed.
- RedBoot "FIS directory" partition scheme layout update
  package/system/mtd/src/fis.c
  Will need to be changed to use the size of the FIS directory partition,
  rather than the partition erasesize.
  The partial erase patchset set erasesize to partition size where
  partition boundary was not major erasesize aligned. This patch sets
  partition erasesize to minor erasesize only where the boundary is
  minor erasesize aligned.

An erasesize_minor variable is added to the MTD struct mtd_info, so that
the mtd level does not need access to the SPI NOR level to calculate the
smallest (minor) universal erasesize. This is set in spi-nor.c device
initialisation.

mtdpart.c is modified to not prevent writing to partitions with a
boundary on a minor eraseblock. As mtd partitions are currently a stack
of structs, the erasesize_minor need to be copied (from parent) to each
partition. MTD partitions use a list in kernel 5.7, but the erasesize
boundary code remains the same.

Signed-off-by: John Thomson <git@johnthomson.fastmail.com.au>