kernel 5.10: mtd: spi-nor: allow use of mixed erasesizes (64K with 4K)
For uniform erase-regions (typical) SPI-NOR devices,
the kernel selects a single erasesize for the device.
This patch allows the kernel to mix any supported erasesizes.

Where the SPI-NOR chip supports it (SECT_4K) a significant speedup
is achieved for large erases (sysupgrade) when 4K erases are required.

This is done by introducting erasesize_minor, and allowing the kernel
to also write to an mtd partition when the partition is on a
minor erasesize boundary.
Minor erasesize will be set for (uniform eraseregion) SPI-NOR devices
with:
 CONFIG_MTD_SPI_NOR_USE_VARIABLE_ERASE=y
which depends on
 # CONFIG_MTD_SPI_NOR_USE_4K_SECTORS is not set

This uses the existing SPI-NOR multiple erase-regions code path for a
uniform erase-region device, which allows an erase request to select and
run a list of erase operations.

Previous users of partial or 4K erase in Openwrt:
- Mikrotik routerboot bootloader configuration
  target/linux/generic/files/drivers/platform/mikrotik/rb_softconfig.c
- RedBoot "FIS directory" partition scheme layout update
  package/system/mtd/src/fis.c

This patch is a partial replacement to the Openwrt
pending-5.4 411 & 412 patches. They allow the kernel to write to
a partition that is not aligned to erasesize by copying, erasing, then
partially restoring original content.

Signed-off-by: John Thomson <git@johnthomson.fastmail.com.au>