kernel: mtd: spi-nor: allow use of mixed erasesizes (E.g. 64K & 4K)
Allow SPI NOR code to be configured to use the multiple erase-regions
code path for a uniform erase-region device. This code path allows an
erase request to select and run a list of erase operations (including 4K
erases where supported by the device, and needed by the request). The
multi erase-regions code path does additional checks before the erase
starts, therefore may take slightly longer before the erase runs on device.

Kernel config options:
Introduces:
CONFIG_MTD_SPI_NOR_USE_VARIABLE_ERASE=y
Used with:
CONFIG_MTD_SPI_NOR_USE_4K_SECTORS=n

Previous users of partial or 4K erase in Openwrt:
- Mikrotik routerboot bootloader configuration
  target/linux/generic/files/drivers/platform/mikrotik/rb_softconfig.c
  Will work with this patch, with the 4K_SECTORS test removed.
- RedBoot "FIS directory" partition scheme layout update
  package/system/mtd/src/fis.c
  Will need to be changed to use the size of the FIS directory partition,
  rather than the partition erasesize.
  The partial erase patchset set erasesize to partition size where
  partition boundary was not major erasesize aligned. This patch sets
  partition erasesize to minor erasesize only where the boundary is
  minor erasesize aligned.

An erasesize_minor variable is added to the MTD struct mtd_info, so that
the mtd level does not need access to the SPI NOR level to calculate the
smallest (minor) universal erasesize. This is set in spi-nor.c device
initialisation.

mtdpart.c is modified to not prevent writing to partitions with a
boundary on a minor eraseblock. As mtd partitions are currently a stack
of structs, the erasesize_minor need to be copied (from parent) to each
partition. MTD partitions use a list in kernel 5.7, but the erasesize
boundary code remains the same.

There appears to be a bug in spi-nor.c for univeral erase-region devices,
in a code path that only this patch uses: A pointer to an address within
a struct is not updated following a memcpy of the struct.

Signed-off-by: John Thomson <git@johnthomson.fastmail.com.au>

---

Replaces Openwrt generic patches:
These cuts are not included in this patch to reduce RFC length.
- partial erase 411 & 412
- 4K limit 470 (non-functional on 5.4)

Any suggestions welcome!

Further discussion here: https://github.com/openwrt/openwrt/pull/3271

WIP: Do not run this without a full NOR backup,
and a means to restore it.
E.g. SOIC8 clip? + single board computer with extra SPI.

Run tested with Mikrotik soft_config writes on ath79 & ramips.
Not tested on Redboot FIS layout update (fis code needs changes).

An earlier patch which modified the default universal erase-region code
path in spi-nor.c was sent RFC to linux-mtd, but received no comments.

The update pointer to memcpy'd address patch has been recently sent RFC
to linux-mtd and maintainers. My intention is that once this gets addressed,
I send through the erasesize_minor and mtdpart changes.

- I am not sure that testing if (wr_alignment_minor) is enough to check
  that the device has an erasesize_minor set?
- I am not sure that the picked erasesize_minor would be universal for
  multi erase-regions devices.

It is my opinion that code requesting an mtd_erase should manage the
whole eraseblock. The partial erase patchset currently manages this in
kernel, and a recently fixed merge bug in it was able to cause data loss
and bootloader corruption on some Mikrotik and Redboot devices. In
kernel 5.7, the mtd partition code is changed and would require a
rewrite of the partial erase, as part_erase is no longer used.

There may be devices that currently use partial erase which do not support 4K
erase. They would lose functionaly with this patch.

If there are any other user of 4K or partial erase, please let me know.

Cheers!