ath79: mikrotik: use RouterBoot initrd for ramdisk
RouterBoot will copy the ELF section "initrd" to memory.
It is copied to PAGE_ALIGN(end of kernel + 8).
Immediately before this address is a header of
u32 "INRD", u32 size_of_initrd.

For the self-extracting kernel:
If the INRD header is found,
copy the header, and initrd to a point after the
decompressed kernel + vmlinux.ld.S appended DTB space (0x100000) +
room for BSS (+0x100000)

Conditionally reduce the xz dictionary size, so that the decompression
dictionary fits in RAM.

Signed-off-by: John Thomson <git@johnthomson.fastmail.com.au>